Time: 3 ms (65.59%) | Memory: 19.1 MB (6.82%) - LeetSync

+++ 150-evaluate-reverse-polish-notation/evaluate-reverse-polish-notation.py
@@ -14,7 +14,7 @@ class Solution:
                 elif tokens[r] == "-":
                     res = b - a
                 elif tokens[r] == "/":
-                    res = b / a
+                    res = int(b / a)
                 elif tokens[r] == "*":
                     res = b * a
                 stack.append(res)
@@ -22,4 +22,4 @@ class Solution:
                 stack.append(tokens[r])
   
 
-        return int(stack[-1])
+        return stack[-1]
